+++ packages/cli/src/util/integration/fetch-marketplace-integrations-list.ts
@@ -15,9 +15,18 @@ export type IntegrationListItem = {
   canInstall?: boolean;
 };
 
-export async function fetchMarketplaceIntegrationsList(client: Client) {
+export async function fetchMarketplaceIntegrationsList(
+  client: Client,
+  categories?: string[]
+) {
+  const params = new URLSearchParams({ integrationType: 'marketplace' });
+  if (categories?.length) {
+    for (const category of categories) {
+      params.append('category', category);
+    }
+  }
   return client.fetch<IntegrationListItem[]>(
-    '/v2/integrations/integrations?integrationType=marketplace',
+    `/v2/integrations/integrations?${params.toString()}`,
     {
       json: true,
     }