+++ packages/client/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @vercel/client
 
+## 17.5.15
+
+### Patch Changes
+
+- Updated dependencies [01e18e8]
+  - @vercel/build-utils@13.30.0
+
+## 17.5.14
+
+### Patch Changes
+
+- Updated dependencies [32a730e]
+  - @vercel/build-utils@13.29.1
+
 ## 17.5.13
 
 ### Patch Changes