+++ packages/fastify/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @vercel/fastify
 
+## 0.1.96
+
+### Patch Changes
+
+- @vercel/node@5.8.17
+
+## 0.1.95
+
+### Patch Changes
+
+- @vercel/node@5.8.16
+
 ## 0.1.94
 
 ### Patch Changes